Reorganize Wear Material detail composables
- Move detail components to dedicated package
- Split components into focused files
