+++ ui-wear-compose-material/src/debug/java/com/github/droibit/oss_licenses/ui/wear/compose/material/internal/detail/OssLicenseDetailScreenPreview.kt
@@ -1,4 +1,4 @@
-package com.github.droibit.oss_licenses.ui.wear.compose.material.internal
+package com.github.droibit.oss_licenses.ui.wear.compose.material.internal.detail
 
 import androidx.compose.runtime.Composable
 import androidx.wear.compose.material.MaterialTheme
