+++ ui-wear-compose-material/src/test/java/com/github/droibit/oss_licenses/ui/wear/compose/material/internal/detail/OssLicenseDetailScreenshotTest.kt
@@ -1,4 +1,4 @@
-package com.github.droibit.oss_licenses.ui.wear.compose.material.internal
+package com.github.droibit.oss_licenses.ui.wear.compose.material.internal.detail
 
 import com.github.droibit.oss_licenses.ui.compose.screenshots.TestFixtures
 import com.github.droibit.oss_licenses.ui.compose.screenshots.WearDevice
